Controller updates

+++ src/backend/OnlyBans.Backend/Controllers/PostController.cs
@@ -1,5 +1,7 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.StaticFiles;
+﻿using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using OnlyBans.Backend.Data;
 using OnlyBans.Backend.Models;
@@ -10,7 +12,7 @@ namespace OnlyBans.Backend.Controllers;
 
 [Route("api/v1/[controller]")]
 [ApiController]
-public class PostController(AppDbContext context) : ControllerBase {
+public class PostController(AppDbContext context, UserManager<User> userManager) : ControllerBase {
     
     [HttpGet]
     public async Task<ActionResult<IEnumerable<UserGetDto>>> GetPosts() {
@@ -27,16 +29,20 @@ public class PostController(AppDbContext context) : ControllerBase {
         return Ok(new PostGetDto(post));
     }
     
+    [Authorize]
     [HttpPost("{id:guid}/like")]
     public async Task<IActionResult> LikePost(Guid id) {
-        var userId = Guid.Empty;
-
+        
+        var user = await userManager.GetUserAsync(User);
+        if (user == null)
+            return Unauthorized(new { message = "You are not logged in!" });
+        
         var post = await context.Posts.FindAsync(id);
         if (post == null)
             return NotFound(new { message = "Post not found" });
         
         var existingLike = await context.UserPostLikes
-            .FirstOrDefaultAsync(l => l.PostId == id && l.UserId == userId);
+            .FirstOrDefaultAsync(l => l.PostId == id && l.UserId == user.Id);
 
         if (existingLike != null) {
             context.UserPostLikes.Remove(existingLike);
@@ -46,11 +52,10 @@ public class PostController(AppDbContext context) : ControllerBase {
         
         var like = new UserPostLike {
             PostId = id,
-            UserId = userId
+            UserId = user.Id
         };
         context.UserPostLikes.Add(like);
         await context.SaveChangesAsync();
         return Ok(new { message = "Post liked successfully" });
     }
-
 }
